+++ lib/src/components/LifecycleProgress/LifecycleTile.tsx
@@ -1,6 +1,6 @@
 
 import type { ILifecycleStages, ILifecycleProgressProps } from '@models/interfaces/ILifecycleProgressProps';
-import { Icon } from '@fluentui/react/lib/Icon';
+import { Checkmark20Regular } from '@fluentui/react-icons';
 import { useContext } from 'react';
 import { CalloutCtx } from './Context';
 
@@ -20,7 +20,7 @@ const LifecycleTile = ({styles, ...props}: LifecycleTileProps<any>) => {
 		<div className={indicatorClass} style={{ order: props.currentIdx, display: props?.hidden ? 'none' : 'flex' }}>
 			{props.completed ? (
 				<span className={styles.stageIndicator}>
-					<Icon styles={{root: {width: '100%' }}} iconName='CheckMark' />
+					<Checkmark20Regular />
 				</span>
 			) : (
 				<span className={styles.stageIndicator}></span>